modeling_pipeline/pipeline_v4/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
ijson>=3.2.0
scikit-learn>=1.3.0
xgboost>=2.0.0
catboost>=1.2.0
optuna>=3.0.0
joblib>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
requests>=2.31.0
tqdm>=4.65.0
psycopg2-binary>=2.9.0

